Remove definition of ValueDeserialize
