+++ merde_core/src/lib.rs
@@ -20,33 +20,12 @@ pub use with_lifetime::WithLifetime;
 mod value;
 pub use value::Value;
 
-mod deserialize;
-pub use deserialize::OwnedValueDeserialize;
-pub use deserialize::ValueDeserialize;
-
 mod deserialize2;
 pub use deserialize2::ArrayStart;
 pub use deserialize2::Deserialize;
 pub use deserialize2::Deserializer;
 pub use deserialize2::Event;
 pub use deserialize2::EventType;
-
-/// Interpret a &[`Value`] as an instance of type `T`. This may involve
-/// more cloning than [`from_value`].
-pub fn from_value_ref<'s, T>(value: &Value<'s>) -> Result<T, MerdeError<'s>>
-where
-    T: ValueDeserialize<'s>,
-{
-    T::from_value_ref(Some(value))
-}
-
-/// Interpret a [`Value`] as an instance of type `T`.
-pub fn from_value<'s, T>(value: Value<'s>) -> Result<T, MerdeError<'s>>
-where
-    T: ValueDeserialize<'s>,
-{
-    T::from_value(Some(value))
-}
 
 rubicon::compatibility_check! {
     ("merde_core_pkg_version", env!("CARGO_PKG_VERSION")),
